+++ wails/main.go
@@ -65,7 +65,6 @@ func main() {
 		Windows: &windows.Options{
 			WebviewIsTransparent: false,
 			WindowIsTranslucent:  false,
-			DisableWindowIcon:    true,
 		},
 		Mac: &mac.Options{
 			OnFileOpen: app.onMacFileOpen,